+++ test/Telegram.Bot.Tests.Integ/Framework/Fixtures/PrivateChatFixture.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading.Tasks;
+using System.Threading.Tasks;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;