+++ CHANGELOG.md
@@ -11,6 +11,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Additional constructors on `InputMediaPhoto` and `InputMediaVideo`
 
+## [14.6.0] - 2018-06-12
+
+### Added
+
+- Property `ParseMode` to requests with a caption
+  - `EditMessageCaptionRequest`
+  - `EditInlineMessageCaptionRequest`
+- Parameter `parseMode` to method `ITelegramBotClient.EditMessageCaptionAsync`
+
+## [14.5.0] - 2018-06-06
+
+### Added
+
+- New enum member `MessageEntityType.PhoneNumber`
+- New enum member `MessageEntityType.Unknown`
+
+### Fixed
+
+- Exception during deserialization of unknown message entity type
+
 ## [14.4.0] - 2018-05-17
 
 ### Changed